Implements Payment service for course enrollment.
Puts proper authorization on the current resources so that the students
can access only the resources that they need and any other resources
require admin authorization.

Uses @PreAuthorize (Method Level security).

Fixes JSON serialization issue with the entities by annotating the
field references with @JsonManagedReference and @JsonBackReference.

Other minor changes.

+++ src/main/java/com/nturbo1/satWebsiteProjBack/service/dto/request/courses/CourseRequestDto.java
@@ -13,7 +13,7 @@ public record CourseRequestDto(
 	String type,
 	Date createdDate,
 	String status,
-	Double price,
+	Integer price,
 	List<CourseSection> sections,
 	List<User> enrolledStudents
 ) {
